Prof. Maikon/requirements.txt
flask~=3.0.2
pillow~=10.2.0
pandas~=2.2.1
numpy~=1.26.4
WTForms~=3.1.2

eventlet
flask
flask_sqlalchemy
flask_bcrypt
flask_socketio
flask_login
flask_wtf
gunicorn
